+++ BB_RPG_DiceSet/Contracts/IModify.cs
@@ -1,0 +1,13 @@
+﻿namespace BB_RPG_DiceSet
+{
+    /// <summary>
+    /// Implement this interface if you want to create ways to modify a Roll and specific game mechanics like:
+    /// AdvantageModifier
+    /// DisadvantageModifier
+    /// etc...
+    /// </summary>
+    public interface IModifier
+    {
+        int Modify(IRollable rollableObject);
+    }
+}