+++ BB_RPG_DiceSet/Rollables/StandardDice.cs
@@ -1,0 +1,37 @@
+﻿namespace BB_RPG_DiceSet
+{
+    public class D4 : Die
+    {
+        public D4() : base(4) { }
+    }
+
+    public class D6 : Die
+    {
+        public D6() : base(6) { }
+    }
+
+    public class D8 : Die
+    {
+        public D8() : base(8) { }
+    }
+
+    public class D10 : Die
+    {
+        public D10() : base(10) { }
+    }
+
+    public class D12 : Die
+    {
+        public D12() : base(12) { }
+    }
+
+    public class D20 : Die
+    {
+        public D20() : base(20) { }
+    }
+
+    public class D100 : Die
+    {
+        public D100() : base(100) { }
+    }
+}